heic2jpg/requirements.txt
Pillow==10.0.0
pillow-heif==0.13.0
tqdm==4.66.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
